% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DSU3L}
\alias{DSU3L}
\title{DSU mapping from EQ-5D-3L to EQ-5D-5L}
\format{
An object of class \code{data.frame} with 2430 rows and 60 columns.
}
\source{
Hernández Alava M, Pudney S, Wailoo A. Estimating the Relationship Between EQ-5D-5L and EQ-5D-3L: Results from a UK Population Study. Pharmacoeconomics. 2023 Feb;41(2):199-207. \doi{10.1007/s40273-022-01218-7}. Epub 2022 Nov 30. PMID: 36449173.

Hernández-Alava M, Pudney S. Econometric modelling of multiple self-reports of health states: The switch from EQ-5D-3L to EQ-5D-5L in evaluating drug therapies for rheumatoid arthritis. J Health Econ. 2017 Sep;55:139-152. \doi{10.1016/j.jhealeco.2017.06.013}. Epub 2017 Jul 4. PMID: 28778350.

\href{https://sheffield.ac.uk/nice-dsu/methods-development/mapping-eq-5d-5l-3l}{NICE DSU mapping website}.
}
\usage{
DSU3L
}
\description{
Data for age and sex based mapping from EQ-5D-3L dimensions or utility index
score to EQ-5D-5L for China, Germany, Japan, Netherlands, South Korea, Spain
and UK.
}
\keyword{datasets}
