% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d.R
\name{eq5d}
\alias{eq5d}
\title{Calculate EQ-5D indices.}
\usage{
eq5d(scores, version, type, country)
}
\arguments{
\item{scores}{numeric or data.frame with names/colnames Mobility, Care, Activity, Pain and Anxiety.}

\item{version}{string of value "3L" or "5L" to indicate instrument version.}

\item{type}{string specifying method type used in deriving value set scores. Options 
are TTO or VAS for EQ-5D-3L and VT for EQ-5D-5L.}

\item{country}{string of value set country name used.}
}
\description{
Calculate EQ-5D indices for EQ-5D-3L and EQ-5D-5L. Available value sets can be seen 
  using the function \code{valuesets}
}
\examples{
eq5d(scores=c(Mobility=1,Care=2,Activity=3,Pain=4,Anxiety=5), 
 country="Indonesia", version="5L")
eq5d(scores=c(Mobility=3,Care=2,Activity=3,Pain=2,Anxiety=3), 
 type="TTO", version="3L", country="Germany")

test.df <- data.frame(
  Mobility=c(1,2,3,4,5), Care=c(1,5,4,3,2),
  Activity=c(1,5,2,3,1), Pain=c(1,3,4,3,4), Anxiety=c(1,2,1,2,1)
  )
eq5d(test.df, country="Canada", version="5L")

}
