% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d3l.R
\name{eq5d3l}
\alias{eq5d3l}
\title{Calculate EQ-5D-3L indices.}
\usage{
eq5d3l(scores, type = "TTO", country = "UK")
}
\arguments{
\item{scores}{numeric with names Mobility, Care, Activity, Pain and Anxiety.}

\item{type}{3L values set type. Either TTO or VAS.}

\item{country}{value set country.}
}
\description{
Calculate indices for EQ-5D-3L value sets. Available value sets can be seen 
  using the function \code{valuesets}
}
\examples{
eq5d3l(scores=c(Mobility=1,Care=2,Activity=3,Pain=1,Anxiety=3), type="VAS", country="UK")
eq5d3l(scores=c(Mobility=3,Care=2,Activity=3,Pain=2,Anxiety=3), type="TTO", country="Germany")

}
