% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d5l.R
\name{eq5d5l}
\alias{eq5d5l}
\title{Calculate EQ-5D-5L indices.}
\usage{
eq5d5l(scores, country = "England")
}
\arguments{
\item{scores}{numeric with names Mobility, Care, Activity, Pain and Anxiety.}

\item{country}{value set country.}
}
\description{
Calculate indices for EQ-5D-5L value sets. Available value sets can be seen 
  using the function \code{valuesets}
}
\examples{
eq5d5l(scores=c(Mobility=1,Care=2,Activity=3,Pain=4,Anxiety=5), country="England")
eq5d5l(scores=c(Mobility=3,Care=2,Activity=5,Pain=2,Anxiety=3), country="Netherlands")

}
