% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{eq5d_theme}
\alias{eq5d_theme}
\title{eq5d plotting theme}
\usage{
eq5d_theme(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base font size, in points.}

\item{base_family}{Base font family.}
}
\value{
A ggplot2 theme.
}
\description{
Default ggplot2 theme matching the academic style used in
Methods for Analysing and Reporting EQ-5D Data (Devlin et al., 2020).
}
