% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5dy.R
\name{eq5dy}
\alias{eq5dy}
\title{Calculate EQ-5D-Y index scores}
\usage{
eq5dy(scores, country = NULL)
}
\arguments{
\item{scores}{numeric with names MO, SC, UA, PD and AD representing
Mobility, Self-care, Usual activities, Pain/discomfort and Anxiety/depression.}

\item{country}{value set country.}
}
\value{
calculated utility index score.
}
\description{
Calculate indices for EQ-5D-Y value sets. Available value sets can be viewed
  using the function \code{valuesets}.
}
\examples{
eq5dy(scores=c(MO=3,SC=3,UA=3,PD=3,AD=3), country="Slovenia")

}
