% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_all_health_states}
\alias{get_all_health_states}
\alias{getHealthStates}
\title{Get all five digit health state scores}
\usage{
get_all_health_states(version)
}
\arguments{
\item{version}{the EQ-5D version. Either 3L or 5L.}
}
\value{
A character vector of five digit health states.
}
\description{
Get all five digit health state scores for either EQ-5D-3L, EQ-5D-5L or
EQ-5D-Y
}
\examples{
get_all_health_states("3L")
get_all_health_states("5L")
get_all_health_states("Y")

}
