% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_dimensions_from_health_states}
\alias{get_dimensions_from_health_states}
\alias{splitHealthStates}
\alias{getDimensionsFromHealthStates}
\title{Get individual dimension scores from their five digit health states}
\usage{
get_dimensions_from_health_states(
  scores,
  ignore.invalid = TRUE,
  version = "5L"
)
}
\arguments{
\item{scores}{a vector of five digit scores}

\item{ignore.invalid}{whether to ignore invalid scores. TRUE returns NA, FALSE throws an 
error.}

\item{version}{3L, 5L or Y. Used for validating scores when ignore.invalid 
is FALSE.}
}
\value{
A data.frame of individual dimension scores.
}
\description{
Get a data.frame of individual dimension scores from their five digit health states.
}
\examples{
get_dimensions_from_health_states(c("12345", "54321"), version="5L")

}
