% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getHealthStates}
\alias{getHealthStates}
\title{Get all five digit health state scores}
\usage{
getHealthStates(version)
}
\arguments{
\item{version}{the EQ-5D version. Either 3L or 5L.}
}
\value{
A character vector of five digit health states.
}
\description{
Get all five digit health state scores for either EQ-5D-3L, EQ-5D-5L or
EQ-5D-Y
}
\examples{
getHealthStates("3L")
getHealthStates("5L")
getHealthStates("Y")

}
