% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getHealthStatesFromDimensions}
\alias{getHealthStatesFromDimensions}
\title{Get five digit health states from dimension scores}
\usage{
getHealthStatesFromDimensions(
  scores,
  version = "5L",
  ignore.invalid = TRUE,
  dimensions = .getDimensionNames()
)
}
\arguments{
\item{scores}{a data.fram containing each dimension in a column}

\item{version}{3L or 5L. Used for validating scores when ignore.invalid 
is FALSE.}

\item{ignore.invalid}{whether to ignore invalid scores. TRUE returns NA, FALSE throws an 
error.}
}
\value{
A character vector of individual dimension scores.
}
\description{
Merge MO, SC, UA, PD and AD dimension scores to get five digit health states.
}
\examples{
getHealthStatesFromDimensions(scores, version="5L")

}
