% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny_eq5d}
\alias{shiny_eq5d}
\title{Launch shiny EQ-5D interface}
\usage{
shiny_eq5d(display.mode = "normal")
}
\arguments{
\item{display.mode}{The display mode to be passed to \link[shiny]{runApp}}
}
\description{
\code{shiny_eq5d} launches a shiny interface for browser based EQ-5D calculations.
}
\examples{
\dontrun{
shiny_eq5d()
shiny_eq5d(display.mode="normal")
}
}
