% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{splitHealthStates}
\alias{splitHealthStates}
\title{Split five digit health states}
\usage{
splitHealthStates(scores, ignore.invalid = TRUE, version = "5L")
}
\arguments{
\item{scores}{a vector of five digit scores}

\item{ignore.invalid}{whether to ignore invalid scores. TRUE returns NA, FALSE throws an 
error.}

\item{version}{3L or 5L. Used for validating scores when ignore.invalid 
is FALSE.}
}
\value{
A data.frame of individual dimension scores.
}
\description{
Split five digit health states into their individual dimension scores.
}
\examples{
splitHealthStates(c("12345", "54321"), version="5L")

}
