% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TTO}
\alias{TTO}
\title{EQ-5D-3L TTO value set data.}
\format{An object of class \code{data.frame} with 17 rows and 8 columns.}
\source{
Szende, A., Oppe, M., & de Charro, F. (2007), Comparative review of Time Trade-Off value sets. 
  In Szende, A., Oppe, M., & Devlin, N. (Ed.), EQ-5D Value Sets: Inventory, Comparative Review
  and User Guide (pp. 27-28). Dordrecht, The Netherlands: Springer.
}
\usage{
TTO
}
\description{
Coefficients for the estimation of the EQ-5D-3L index values based on TTO valuation studies
  for Denmark, Germany, Japan, Netherlands, Spain, UK, USA and Zimbabwe.
}
\keyword{datasets}
