% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d.R
\name{valuesets}
\alias{valuesets}
\title{Get the available EQ-5D value sets}
\usage{
valuesets()
}
\value{
A data.frame countaining the EQ-5D version, value set type and country
}
\description{
\code{valuesets} returns a data.frame of the available EQ-5D value sets
    in the \code{eq5d} package.
}
\examples{
valuesets()
}
