% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VT}
\alias{VT}
\title{EQ-5D-5L VT value set data}
\format{
An object of class \code{data.frame} with 37 rows and 50 columns.
}
\source{
\itemize{
\item{\strong{Australia}: Norman R, Mulhern B, Lancsar E, Lorgelly P, Ratcliffe J, Street D, Viney R. The Use of a Discrete Choice Experiment Including Both Duration and Dead for the Development of an EQ-5D-5L Value Set for Australia. Pharmacoeconomics. 2023 Jan 31. \doi{10.1007/s40273-023-01243-0}. Epub ahead of print. PMID: 36720793.}
\item{\strong{Belgium}: Bouckaert N, Cleemput I, Devriese S, Gerkens S. An EQ-5D-5L Value Set for Belgium. Pharmacoecon Open. 2022 Aug 4. \doi{10.1007/s41669-022-00353-3}. Epub ahead of print. PMID: 35927410.}
\item{\strong{Canada}: Xie F, Pullenayegum E, Gaebel K, Bansback N, Bryan S, Ohinmaa A, Poissant L, Johnson JA; Canadian EQ-5D-5L Valuation Study Group. A Time Trade-off-derived Value Set of the EQ-5D-5L for Canada. Med Care. 2016 Jan;54(1):98-105. \doi{10.1097/MLR.0000000000000447}. PMID: 26492214.}
\item{\strong{China}: Luo N, Liu G, Li M, Guan H, Jin X, Rand-Hendriksen K. Estimating an EQ-5D-5L Value Set for China. Value Health. 2017 Apr;20(4):662-669. \doi{10.1016/j.jval.2016.11.016}. Epub 2017 Feb 9. PMID: 28408009.}
\item{\strong{Denmark}: Jensen CE, Sørensen SS, Gudex C, Jensen MB, Pedersen KM, Ehlers LH. The Danish EQ-5D-5L Value Set: A Hybrid Model Using cTTO and DCE Data. Appl Health Econ Health Policy. 2021 Feb 2. \doi{10.1007/s40258-021-00639-3}. Epub ahead of print. PMID: 33527304.}
\item{\strong{Egypt}: Al Shabasy S, Abbassi M, Finch A, Roudijk B, Baines D, Farid S. The EQ-5D-5L Valuation Study in Egypt. Pharmacoeconomics. 2021 Nov 17:1–15. \doi{10.1007/s40273-021-01100-y}. Epub ahead of print. PMID: 34786590.}
\item{\strong{England}: Devlin NJ, Shah KK, Feng Y, Mulhern B, van Hout B. Valuing health-related quality of life: An EQ-5D-5L value set for England. Health Econ. 2018 Jan;27(1):7-22. \doi{10.1002/hec.3564}. Epub 2017 Aug 22. PMID: 28833869.}
\item{\strong{Ethiopia}: Welie AG, Gebretekle GB, Stolk E, Mukuria C, Krahn MD, Enquoselassie F, Fenta TG. Valuing Health State: An EQ-5D-5L Value Set for Ethiopians. Value Health Reg Issues. 2019 Nov 1;22:7-14. \doi{10.1016/j.vhri.2019.08.475}. PMID: 31683254.}
\item{\strong{France}: Andrade LF, Ludwig K, Goni JMR, Oppe M, de Pouvourville G. A French Value Set for the EQ-5D-5L. Pharmacoeconomics. 2020 Jan 8. \doi{10.1007/s40273-019-00876-4}. PMID: 31912325.}
\item{\strong{Germany}: Ludwig K, Graf von der Schulenburg JM, Greiner W. German Value Set for the EQ-5D-5L. Pharmacoeconomics. 2018 Feb;36(6):663-674. \doi{10.1007/s40273-018-0615-8}. PMID: 29460066.}
\item{\strong{Ghana}: Addo R, Mulhern B, Norman R, Owusu R, Viney R, Nonvignon J. An EQ-5D-5L Value Set for Ghana Using an Adapted EuroQol Valuation Technology Protocol. Value Health Reg Issues. 2024 Sep 4;45:101045. \doi{10.1016/j.vhri.2024.101045}. Epub ahead of print. PMID: 39236574.}
\item{\strong{HongKong}: Wong ELY, Ramos-Goñi JM, Cheung AWL, Wong AYK, Rivero-Arias O. Assessing the Use of a Feedback Module to Model EQ-5D-5L Health States Values in Hong Kong. Patient. 2018 Apr;11(2):235-247. \doi{10.1007/s40271-017-0278-0}. PMID: 29019161.}
\item{\strong{Hungary}: Rencz F, Brodszky V, Gulácsi L, Golicki D, Ruzsa G, Pickard AS, Law EH, Péntek M. Parallel Valuation of the EQ-5D-3L and EQ-5D-5L by Time Trade-Off in Hungary. Value Health. 2020 Sep;23(9):1235-1245. \doi{10.1016/j.jval.2020.03.019}. Epub 2020 Aug 12. PMID: 32940242.}
\item{\strong{India}: Jyani G, Sharma A, Prinja S, Kar SS, Trivedi M, Patro BK, Goyal A, Purba FD, Finch AP, Rajsekar K, Raman S, Stolk E, Kaur M. Development of an EQ-5D Value Set for India Using an Extended Design (DEVINE) Study: The Indian 5-Level Version EQ-5D Value Set. Value Health. 2022 Jul;25(7):1218-1226. \doi{10.1016/j.jval.2021.11.1370}. Epub 2022 Jan 5. PMID: 35779943.}
\item{\strong{Indonesia}: Purba FD, Hunfeld JAM, Iskandarsyah A, Fitriana TS, Sadarjoen SS, Ramos-Goñi JM, Passchier J, Busschbach JJ. The Indonesian EQ-5D-5L Value Set. PharmacoEconomics. 2017 Nov;35(11)1153-1165. \doi{10.1007/s40273-017-0538-9}. PMID: 28695543.}
\item{\strong{Iran}: Afshari S, Daroudi R, Goudarzi R, Mahboub-Ahari A, Yaseri M, Sari AA, Ameri H, Bahariniya S, Oliaei-Manesh A, Kalavani K, Zare Z, Hasannezhad E, Mirzaei M, Amiri Z. A national survey of Iranian general population to estimate a value set for the EQ-5D-5L. Qual Life Res. 2023 Mar 10. \doi{10.1007/s11136-023-03378-1}. Epub ahead of print. PMID: 36897530.}
\item{\strong{Ireland}: Hobbins A, Barry L, Kelleher D, Shah K, Devlin N, Ramos Goñi JM, O’Neill C. Utility Values for Health States in Ireland: A Value Set for the EQ-5D-5L. PharmacoEconomics. 2018 Nov;36(11):1345-1353. \doi{10.1007/s40273-018-0690-x}. PMID: 30051267.}
\item{\strong{Italy}: Finch AP, Meregaglia M, Ciani O, Roudijk B, Jommi C. An EQ-5D-5L value set for Italy using videoconferencing interviews and feasibility of a new mode of administration. Soc Sci Med. 2022 Jan;292:114519. \doi{10.1016/j.socscimed.2021.114519}. Epub 2021 Oct 28. PMID: 34736804.}
\item{\strong{Japan}: Shiroiwa T, Ikeda S, Noto S, Igarashi A, Fukuda T, Saito S, Shimozuma K. Comparison of Value Set Based on DCE and/or TTO Data: Scoring for EQ-5D-5L Health States in Japan. Value Health. 2016 Jul-Aug;19(5):648-54. \doi{10.1016/j.jval.2016.03.1834}. Epub 2016 Apr 26. PMID: 27565282.}
\item{\strong{Malaysia}: Shafie AA; Vasan Thakumar A; Lim CJ;Luo N; Rand-Hendriksen K; Yusof FA. EQ-5D-5L Valuation for the Malaysian Population. PharmacoEconomics. 2019 May;37(5):715-725. \doi{10.1007/s40273-018-0758-7}. PMID: 30535779.}
\item{\strong{Mexico}: Gutierrez-Delgado C, Galindo-Suárez RM, Cruz-Santiago C, Shah K, Papadimitropoulos M, Feng Y, Zamora B, Devlin N. EQ-5D-5L Health-State Values for the Mexican Population. Appl Health Econ Health Policy. 2021 Nov;19(6):905-914. \doi{10.1007/s40258-021-00658-0}. Epub 2021 Jun 26. PMID: 34173957.}
\item{\strong{Morocco}: Azizi A, Boutib A, Achak D, Purba FD, Rencz F, Saad E, Hilali A, Ahid S, Nejjari C, Stolk EA, Roudijk B, Youlyouz-Marfak I, Marfak A.Valuing health-related quality of life: an EQ-5D-5L value set for Morocco. Qual Life Res. 2025 Feb 28. \doi{10.1007/s11136-025-03930-1}. Online ahead of print. PMID: 40019677.}
\item{\strong{Netherlands}: Versteegh MM, Vermeulen KM, Evers SM, de Wit GA, Prenger R, Stolk EA. Dutch Tariff for the Five-Level Version of EQ-5D. Value in Health. 2016 Jun;19(4):343-52.  \doi{10.1016/j.jval.2016.01.003}. PMID: 27325326.}
\item{\strong{New Zealand}: Sullivan T, Hansen P, Ombler F, Derrett S, Devlin N. A new tool for creating personal and social EQ-5D-5L value sets, including valuing 'dead'. Soc Sci Med. 2020 Feb;246:112707. \doi{10.1016/j.socscimed.2019.112707}. Epub 2019 Nov 30. PMID: 31945596.}
\item{\strong{Nigeria}: Yusuf AH, Ardo BU, Thavorncharoensap M, Roudijk B, Purba FD, Yang Z, Liao M, Chaikledkaew U, Youngkong S, Thakkinstian A, Agada-Amade YA, Amole TG, Sambo MN, Ohiri K. The EQ-5D-5L valuation study in Nigeria. Qual Life Res. 2026 Jun 15;35(8):206. \doi{10.1007/s11136-026-04319-4}. PMID: 42295458.}
\item{\strong{Norway}: Garratt AM, Stavem K, Shaw JW, Rand K. EQ-5D-5L value set for Norway: a hybrid model using cTTO and DCE data. Qual Life Res. 2024 Nov 20. \doi{10.1007/s11136-024-03837-3}. Epub ahead of print. PMID: 39565555.}
\item{\strong{Peru} Augustovski F, Belizán M, Gibbons L, Reyes N, Stolk E, Craig BM, Tejada RA. Peruvian Valuation of the EQ-5D-5L: A Direct Comparison of Time Trade-Off and Discrete Choice Experiments. Value Health. 2020;23(7):880-888. \doi{10.1016/j.jval.2020.05.004}. PMID: 32762989.}
\item{\strong{Philippines} Miguel RTD, Rivera AS, Cheng KJG, Rand K, Purba FD, Luo N, Zarsuelo MA, Genuino-Marfori AJ, Florentino-Fariñas I, Guerrero AM, Lam HY. Estimating the EQ-5D-5L value set for the Philippines. Qual Life Res. 2022 May 9. \doi{10.1007/s11136-022-03143-w}. PMID: 35532835.}
\item{\strong{Poland} Golicki D, Jakubczyk M, Niewada M, Wrona W, Busschbach JJ. Valuation of EQ-5D health states in Poland: first TTO-based social value set in Central and Eastern Europe. Value Health. 2010 Mar-Apr;13(2):289-97. \doi{10.1111/j.1524-4733.2009.00596.x}. PMID: 19744296.}
\item{\strong{Portugal} Ferreira PL, Antunes P, Ferreira LN, Pereira LN, Ramos-Goñi JM. A hybrid modelling approach for eliciting health state preferences: the Portuguese EQ-5D-5L value set. Qual Life Res. 2019 Jun 14. \doi{10.1007/s11136-019-02226-5}. PMID: 31201730.}
\item{\strong{Romania} Olariu E, Mohammed W, Oluboyede Y, Caplescu R, Niculescu-Aron IG, Paveliu MS, Vale L. EQ-5D-5L: a value set for Romania. Eur J Health Econ. 2022 Jun 10. \doi{10.1007/s10198-022-01481-7}. Epub ahead of print. PMID: 35688994.}
\item{\strong{Saudi Arabia} Al-Jedai A, Almudaiheem H, Al-Salamah T, Aldosari M, Almutairi AR, Almogbel Y, AlRuthia Y, Althemery AU, Alluhidan M, Roudijk B, Purba FD, Awad N, O'jeil R. Valuation of EQ-5D-5L in the Kingdom of Saudi Arabia: A national representative study. Value Health. 2024 Feb 9:S1098-3015(24)00047-0. \doi{10.1016/j.jval.2024.01.017}. PMID: 38342365.}
\item{\strong{Singapore} Luo N, Vasan Thakumar A, Cheng LJ, Yang Z, Rand K, Cheung YB, Thumboo J. Developing an EQ-5D-5L Value Set for Singapore. Pharmacoeconomics. 2025 Aug 29. \doi{10.1007/s40273-025-01519-7}. Epub ahead of print. PMID: 40880001.}
\item{\strong{Slovenia} Prevolnik Rupel V, Ogorevc M. EQ-5D-5L Value Set for Slovenia. Pharmacoeconomics. 2023 Jun 21. \doi{10.1007/s40273-023-01280-9}. Epub ahead of print. PMID: 37341959.}
\item{\strong{South Korea} Kim SH, Ahn J, Ock M, Shin S, Park J, Luo N, Jo MW. The EQ-5D-5L valuation study in Korea. Qual Life Res. 2016 Jul;25(7):1845-52. \doi{10.1007/s11136-015-1205-2}. PMID: 26961008.}
\item{\strong{Spain}: Ramos-Goñi JM, Craig B, Oppe M, Ramallo-Fariña Y, Pinto-Prades JL, Luo N, Rivero-Arias O. Handling data quality issues to estimate the Spanish EQ-5D-5L Value Set using a hybrid interval regression approach.  Value in Health 2018 May;21(5):596-604. \doi{10.1016/j.jval.2017.10.023}. PMID: 29753358.}
\item{\strong{Sweden (2020)}: Burström K, Teni FS, Gerdtham UG, Leidl R, Helgesson G, Rolfson O, Henriksson M. Experience-Based Swedish TTO and VAS Value Sets for EQ-5D-5L Health States. Pharmacoeconomics. 2020 Apr 20. \doi{10.1007/s40273-020-00905-7}. PMID: 32307663.}
\item{\strong{Sweden (2022)}: Sun S, Chuang LH, Sahlén KG, Lindholm L, Norström F. Estimating a social value set for EQ-5D-5L in Sweden. Health Qual Life Outcomes. 2022 Dec 23;20(1):167. \doi{10.1186/s12955-022-02083-w}. PMID: 36564844.}
\item{\strong{Taiwan}: Lin HW, Li CI, Lin FJ, Chang JY, Gau CS, Luo N, Pickard AS, Ramos Goñi JM, Tang CH, Hsu CN. Valuation of the EQ-5D-5L in Taiwan. PLoS One. 2018; 13(12):: e0209344. \doi{10.1371/journal.pone.0209344}. PMID: 30586400.}
\item{\strong{Thailand} Pattanaphesaj J, Thavorncharoensap M, Ramos-Goñi JM, Tongsiri S, Ingsrisawang L, Teerawattananon Y. The EQ-5D-5L Valuation study in Thailand. Expert Review of Pharmacoeconomics & Outcomes Research. 2018 Oct;18(5):551-558. \doi{10.1080/14737167.2018.1494574}. PMID: 29958008.}
\item{\strong{Trinidad_and_Tobago} Bailey H, Jonker MF, Pullenayegum E, Rencz F, Roudijk B. The EQ-5D-5L valuation study for Trinidad and Tobago. Health Qual Life Outcomes. 2024 Jul 2;22(1):51. \doi{10.1186/s12955-024-02266-7}. PMID: 38956543.}
\item{\strong{UAE} Al Sayah F, Roudijk B, El Sadig M, Al Mannaei A, Farghaly MN, Dallal S, Kaddoura R, Metni M, Elbarazi I, Kharroubi SA. A Value Set for the EQ-5D-5L for United Arab Emirates. Value Health. 2025 Jan 27:S1098-3015(25)00021-X. \doi{10.1016/j.jval.2025.01.003}. Epub ahead of print. PMID: 39880198.}
\item{\strong{Uganda} Yang F, Katumba KR, Roudijk B, Yang Z, Revill P, Griffin S, Ochanda PN, Lamorde M, Greco G, Seeley J, Sculpher M. Developing the EQ-5D-5L Value Set for Uganda Using the 'Lite' Protocol. Pharmacoeconomics. 2021 Nov 29:1–13. \doi{10.1007/s40273-021-01101-x}. PMID: 34841471.}
\item{\strong{UK} Rowen D, Mukuria C, Bray N, Carlton J, Longworth L, Meads D, Oluboyede Y, O'Neill C, Yang Y. A UK value set for the EQ-5D-5L. Value Health. 2026 Mar 24:S1098-3015(26)00110-5. \doi{10.1016/j.jval.2026.03.008}. Epub ahead of print. PMID: 41887403.}
\item{\strong{Uruguay}: Augustovski F, Rey-Ares L, Irazola V, Garay OU, Gianneo O, Fernández G, Morales M, Gibbons L, Ramos-Goñi JM. An EQ-5D-5L value set based on Uruguayan population preferences. Qual Life Res. 2016 Feb;25(2):323-33. \doi{10.1007/s11136-015-1086-4}. PMID: 26242249.}
\item{\strong{USA}: Pickard AS, Law EH, Jiang R, Pullenayegum E, Shaw JW, Xie F, Oppe M, Boye KS, Chapman RH, Gong CL, Balch A, Busschbach JJV. United States Valuation of EQ-5D-5L Health States Using an International Protocol. Value in Health. 2019 Aug;22(8):931-941. \doi{10.1016/j.jval.2019.02.009}. PMID: 31426935.}
\item{\strong{Vietnam}: Mai VQ, Sun S, Minh HV, Luo N, Giang KB, Lindholm L, Sahlen KG. An EQ-5D-5L Value Set for Vietnam. Qual Life Res. 2020;29(7):1923-1933. \doi{10.1007/s11136-020-02469-7}. PMID: 32221805.}
\item{\strong{WePP}: Olsen JA, Lamu AN, Cairns J. In search of a common currency: A comparison of seven EQ-5D-5L value sets. Health Econ. 2018 Jan;27(1):39-49. \doi{10.1002/hec.3606}. Epub 2017 Oct 24. PMID: 29063633.}
}
}
\usage{
VT
}
\description{
EQ-5D-5L VT value set calculation data for Australia, Belgium, Canada, China, 
Denmark, Egypt, England, Ethiopia, France, Germany, Ghana, HongKong, Hungary,
India, Indonesia, Iran, Ireland, Italy, Japan, Malaysia, Mexico, Morocco, 
Netherlands, NewZealand, Norway, Peru, Philippines, Poland, Portugal, 
Romania, SaudiArabia, Slovenia, SouthKorea, Spain, Sweden, Taiwan, Thailand, 
Uganda, UAE, Uruguay, USA, Vietnam and Western Preference Pattern (WePP).
}
\keyword{datasets}
