% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y}
\alias{Y}
\title{EQ-5D-Y value set data}
\format{
An object of class \code{data.frame} with 11 rows and 1 columns.
}
\source{
\enumerate{
    \item{\strong{Slovenia}: Prevolnik Rupel V, Ogorevc M; IMPACT HTA HRQoL Group. EQ-5D-Y Value Set for Slovenia. Pharmacoeconomics. 2021 Feb 10. doi: 10.1007/s40273-020-00994-4. Epub ahead of print. \href{https://pubmed.ncbi.nlm.nih.gov/33565048/}{PubMed}}
  }
}
\usage{
Y
}
\description{
EQ-5D-Y value set calculation data for Slovenia.
}
\keyword{datasets}
