% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y3L}
\alias{Y3L}
\title{EQ-5D-Y-3L value set data}
\format{
An object of class \code{data.frame} with 14 rows and 10 columns.
}
\source{
\itemize{
    \item{\strong{Belgium}: Dewilde S, Roudijk B, Tollenaar NH, Ramos-Goñi JM. An EQ-5D-Y-3L Value Set for Belgium. Pharmacoeconomics. 2022 Nov 1:1–12. \doi{10.1007/s40273-022-01187-x}. Epub ahead of print.  PMID: 36316544.}
    \item{\strong{Brazil}: Espirito Santo CM, Miyamoto GC, Santos VS, Ben ÂJ, Finch AP, Roudijk B, de Jesus-Moraleida FR, Stein AT, Santos M, Yamato TP. Estimating an EQ-5D-Y-3L Value Set for Brazil. Pharmacoeconomics. 2024 Jul 2. \doi{10.1007/s40273-024-01404-9}. Epub ahead of print. PMID: 38954389.}
    \item{\strong{China}: Yang Z, Jiang J, Wang P, Jin X, Wu J, Fang Y, Feng D, Xi X, Li S, Jing M, Zheng B, Huang W, Luo N. Estimating an EQ-5D-Y-3L Value Set for China. Pharmacoeconomics. 2022 Nov 18. \doi{10.1007/s40273-022-01216-9}. Epub ahead of print. PMID: 36396878.}
    \item{\strong{Germany}: Kreimeier S, Mott D, Ludwig K, Greiner W; IMPACT HTA HRQoL Group. EQ-5D-Y Value Set for Germany. Pharmacoeconomics. 2022 May 23:1–13. \doi{10.1007/s40273-022-01143-9}. Epub ahead of print. PMID: 35604633.}
    \item{\strong{Hungary}: Rencz F, Ruzsa G, Bató A, Yang Z, Finch AP, Brodszky V. Value Set for the EQ-5D-Y-3L in Hungary. Pharmacoeconomics. 2022 Sep 20:1–11. \doi{10.1007/s40273-022-01190-2}. Epub ahead of print. PMID: 36123448.}
    \item{\strong{Indonesia}: Fitriana TS, Roudijk B, Purba FD, Busschbach JJV, Stolk E. Estimating an EQ-5D-Y-3L Value Set for Indonesia by Mapping the DCE onto TTO Values. Pharmacoeconomics. 2022 Nov 9. \doi{10.1007/s40273-022-01210-1}. Epub ahead of print.  PMID: 36348155.} 
    \item{\strong{Japan}: Shiroiwa T, Ikeda S, Noto S, Fukuda T, Stolk E. Valuation Survey of EQ-5D-Y Based on the International Common Protocol: Development of a Value Set in Japan. Med Decis Making. 2021 Mar 23:272989X211001859. \doi{10.1177/0272989X211001859}. Epub ahead of print. PMID: 33754886.}
    \item{\strong{Netherlands}:Roudijk B, Sajjad A, Essers B, Lipman S, Stalmeier P, Finch AP. A Value Set for the EQ-5D-Y-3L in the Netherlands. Pharmacoeconomics. 2022 Oct 10:1–11. \doi{10.1007/s40273-022-01192-0}. Epub ahead of print. PMID: 36216977.}
    \item{\strong{Slovenia}: Prevolnik Rupel V, Ogorevc M; IMPACT HTA HRQoL Group. EQ-5D-Y Value Set for Slovenia. Pharmacoeconomics. 2021 Feb 10. \doi{10.1007/s40273-020-00994-4}. Epub ahead of print. PMID: 33565048.}
    \item{\strong{Spain}: Ramos-Goñi JM, Oppe M, Estévez-Carrillo A, Rivero-Arias O; IMPACT HTA HRQoL Group. Accounting for Unobservable Preference Heterogeneity and Evaluating Alternative Anchoring Approaches to Estimate Country-Specific EQ-5D-Y Value Sets: A Case Study Using Spanish Preference Data. Value Health. 2022 May;25(5):835-843. \doi{10.1016/j.jval.2021.10.013}. Epub 2021 Dec 6. PMID: 35500952..}
  }
}
\usage{
Y3L
}
\description{
EQ-5D-Y-3L value set calculation data for Belgium, Brazil, China, Germany, Hungary, 
Indonesia, Japan, Netherlands, Slovenia and Spain.
}
\keyword{datasets}
