% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DSU3L}
\alias{DSU3L}
\title{DSU EQ-5D-3L to EQ-5D-5L}
\format{
An object of class \code{data.frame} with 2430 rows and 22 columns.
}
\usage{
DSU3L
}
\description{
DSU EQ-5D-3L to EQ-5D-5L
}
\keyword{datasets}
