% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DSU5L}
\alias{DSU5L}
\title{DSU EQ-5D-5L to EQ-5D-3L}
\format{
An object of class \code{data.frame} with 31250 rows and 22 columns.
}
\usage{
DSU5L
}
\description{
DSU EQ-5D-5L to EQ-5D-3L
}
\keyword{datasets}
