% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d-package.R
\docType{package}
\name{eq5d-package}
\alias{eq5d-package}
\title{eq5d: Methods for Analysing 'EQ-5D' Data and Calculating 'EQ-5D' Index Scores}
\description{
EQ-5D is a widely used health-related quality-of-life instrument developed by the EuroQol Group and used in the clinical and economic evaluation of health care. Health is described using five dimensions (mobility, self-care, usual activities, pain/discomfort, and anxiety/depression) rated on either a three-level (EQ-5D-3L and EQ-5D-Y-3L) or five-level (EQ-5D-5L) scale. Responses can be reported as EQ-5D health states or converted to utility index scores using country-specific value sets. The package provides methods for the valuation, reporting and analysis of EQ-5D data. Utility index scores can be calculated for EQ-5D-3L, EQ-5D-5L and EQ-5D-Y-3L data using a wide range of value sets and mapping approaches. Functionality is also provided for descriptive-system reporting, severity and distributional summaries, informativity measures, health-state distribution analysis, longitudinal change analysis, probability of superiority analysis and Health Profile Grid visualisation. Methods described in Devlin et al. (2020) \doi{10.1007/978-3-030-47622-9} are implemented where appropriate. A companion 'Shiny' application is included for interactive analysis and visualisation of EQ-5D datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fragla/eq5d}
  \item Report bugs at \url{https://github.com/fragla/eq5d/issues}
}

}
\author{
\strong{Maintainer}: Fraser Morton \email{fraser.morton@glasgow.ac.uk}

Authors:
\itemize{
  \item Jagtar Singh Nijjar
}

}
\keyword{internal}
