% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d-package.R
\docType{package}
\name{eq5d-package}
\alias{eq5d-package}
\alias{_PACKAGE}
\title{eq5d: An Implementation of Functions for Calculating 'EQ-5D' Indices}
\description{
An implementation of functions for calculating EQ-5D indices in R. The 
    package can calculate indices for both EQ-5D-3L and EQ-5D-5L value sets. Additionally, a 
    shiny web tool is included to enable the calculation of multiple EQ-5D index values via a 
    web browser using EQ-5D dimension scores stored in CSV or xlsx files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fragla/eq5d}
  \item Report bugs at \url{https://github.com/fragla/eq5d/issues}
}

}
\author{
\strong{Maintainer}: Fraser Morton \email{fraser.morton@glasgow.ac.uk}

Authors:
\itemize{
  \item Jagtar Singh Nijjar \email{jsn33@medschl.cam.ac.uk}
}

}
\keyword{internal}
