% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d5l.R
\name{eq5d5l}
\alias{eq5d5l}
\title{Calculate EQ-5D-5L index scores}
\usage{
eq5d5l(scores, country = "England")
}
\arguments{
\item{scores}{numeric with names MO, SC, UA, PD and AD representing
Mobility, Self-care, Usual activities, Pain/discomfort and Anxiety/depression.}

\item{country}{value set country.}
}
\value{
calculated utility index score.
}
\description{
Calculate indices for EQ-5D-5L value sets. Available value sets can be viewed 
  using the function \code{valuesets}.
}
\examples{
eq5d5l(scores=c(MO=1,SC=2,UA=3,PD=4,AD=5), country="England")
eq5d5l(scores=c(MO=3,SC=2,UA=5,PD=2,AD=3), country="Netherlands")

}
