% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5dcf.R
\name{eq5dcf}
\alias{eq5dcf}
\title{Calculate the cumulative frequency profile of an EQ-5D dataset}
\usage{
eq5dcf(data, version, ignore.invalid, proportions, digits, ...)
}
\arguments{
\item{data}{data.frame with names MO, SC, UA, PD and AD representing
Mobility, Self-care, Usual activities, Pain/discomfort and Anxiety/depression.}

\item{version}{string of value "3L" or "5L" to indicate instrument version.}

\item{ignore.invalid}{booloean whether to ignore invalid scores. TRUE returns NA, FALSE throws an 
error.}

\item{proportions}{boolean whether to include proportion data columns Proportions and
CumulativeProp. Default is FALSE.}

\item{digits}{numeric specifying the number of decimal places for percentages. Defaults to 1.}

\item{...}{character vector, specifying "dimensions" column names. Defaults 
are "MO", "SC", "UA", "PD" and "AD".}
}
\value{
a data.frame or list of data.frames of counts/percentages. Columns 
contain dimensions names and rows the EQ-5D score.
}
\description{
Calculate the frequency, percentage, cumulative frequency and cumulative 
percentage for each profile in an EQ-5D dataset.
}
\examples{
dat <- read.csv(system.file("extdata", "eq5d3l_example.csv", package="eq5d"))
eq5dcf(dat, "3L")

}
