% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5drcw.R
\name{eq5drcw}
\alias{eq5drcw}
\title{Calculate EQ-5D-3L reverse crosswalk index scores}
\usage{
eq5drcw(scores, country = "UK")
}
\arguments{
\item{scores}{numeric with names MO, SC, UA, PD and AD representing
Mobility, Self-care, Usual activities, Pain/discomfort and Anxiety/depression.}

\item{country}{value set country.}
}
\value{
calculated utility index score.
}
\description{
Calculate indices for EQ-5D-3L indices by mapping them onto EQ-5D-5L value sets.
Available value sets can be viewed using the function \code{valuesets}.
}
\examples{
eq5drcw(scores=c(MO=1,SC=2,UA=3,PD=2,AD=1), country="Netherlands")
eq5drcw(scores=c(MO=3,SC=3,UA=3,PD=3,AD=3), country="Germany")

}
