% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RCWVH}
\alias{RCWVH}
\title{EQ-5D-3L Reverse Crosswalk data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 243 rows and 47 columns.
}
\source{
\doi{10.1016/j.jval.2021.03.009}
}
\usage{
RCWVH
}
\description{
Reverse Crosswalk index value table to calculate EQ-5D-5L indices from EQ-5D-3L data
  using the van Hout et al (2021) method.
}
\keyword{datasets}
