% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TTO}
\alias{TTO}
\title{EQ-5D-3L TTO value set data}
\format{
An object of class \code{data.frame} with 63 rows and 32 columns.
}
\source{
Szende, A., Oppe, M., & de Charro, F. (2007), Comparative review of Time Trade-Off value sets.
  In Szende, A., Oppe, M., & Devlin, N. (Ed.), EQ-5D Value Sets: Inventory, Comparative Review
  and User Guide (pp. 27-28). Dordrecht, The Netherlands: Springer.

Janssen, B., Szende, A., & Ramos-Goñi JM. (2014), Data and Methods.
  Szende, A., Janssen, B., & Cabasés, J. (Ed.), In Self-Reported Population Health: An
  International Perspective based on EQ-5D (p 13). Dordrecht, The Netherlands: Springer.
\itemize{
\item{\strong{Argentina}: Augustovski FA, Irazola VE, Velazquez AP, Gibbons L, Craig BM. Argentine valuation of the EQ-5D health states. Value Health. 2009 Jun;12(4):587-96. \doi{10.1111/j.1524-4733.2008.00468.x}. Epub 2008 Nov 12. PMID: 19900257.}
\item{\strong{Australia}: Viney R, Norman R, King MT, Cronin P, Street DJ, Knox S, Ratcliffe J. Time trade-off derived EQ-5D weights for Australia. Value Health. 2011 Sep-Oct;14(6):928-36. \doi{10.1016/j.jval.2011.04.009}. PMID: 21914515.}
\item{\strong{Bermuda}: Bailey H, Roudijk B, Brathwaite R. The EQ-5D-3L valuation study for Bermuda: using an on-line EQ-VT protocol. Eur J Health Econ. 2024 Jul 9. \doi{10.1007/s10198-024-01701-2}. Epub ahead of print. PMID: 38982011.}
\item{\strong{Brazil}: Viegas Andrade M, Noronha K, Kind P, Maia AC, Miranda de Menezes R, De Barros Reis C, Nepomuceno Souza M, Martins D, Gomes L, Nichele D, Calazans J, Mascarenhas T, Carvalho L, Lins C. Societal Preferences for EQ-5D Health States from a Brazilian Population Survey. Value Health Reg Issues. 2013 Dec;2(3):405-412. doi: 10.1016/j.vhri.2013.01.009. Epub 2013 Mar 13. Erratum in: Value Health Reg Issues. 2016 Dec;11:85-87. \doi{10.1016/j.vhri.2016.12.001}. PMID: 29702778.}
\item{\strong{Canada}: Bansback N, Tsuchiya A, Brazier J, Anis A. Canadian valuation of EQ-5D health states: preliminary value set and considerations for future valuation studies. PLoS One. 2012;7(2):e31115. \doi{10.1371/journal.pone.0031115}. Epub 2012 Feb 6. PMID: 22328929.}
\item{\strong{Chile}: Zarate V, Kind P, Valenzuela P, Vignau A, Olivares-Tirado P, Munoz A. Social valuation of EQ-5D health states: the Chilean case. Value Health. 2011 Dec;14(8):1135-41. \doi{10.1016/j.jval.2011.09.002}. Epub 2011 Nov 6. PMID: 22152184.}
\item{\strong{China}: Liu GG, Wu H, Li M, Gao C, Luo N. Chinese time trade-off values for EQ-5D health states. Value Health. 2014 Jul;17(5):597-604. \doi{10.1016/j.jval.2014.05.007}. Epub 2014 Jul 23. PMID: 25128053.}
\item{\strong{Denmark}: Wittrup-Jensen KU, Lauridsen J, Gudex C, Pedersen KM. Generation of a Danish TTO value set for EQ-5D health states. Scand J Public Health. 2009 Jul;37(5):459-66. \doi{10.1177/1403494809105287}. Epub 2009 May 1. PMID: 19411320.}
\item{\strong{Ecuador}: Lucio R, Flores V,  Granja M, Mata G. Resultados de la encuesta de valoración social de los estados de salud de lAños de vida ajustados por calidad (QALY'S). 2019. \href{https://www.researchgate.net/publication/335840590_QALYs_FINAL_Ecuadordig}{Link}}
\item{\strong{France}: Chevalier J, de Pouvourville G. Valuing EQ-5D using time trade-off in France. Eur J Health Econ. 2013 Feb;14(1):57-66. \doi{10.1007/s10198-011-0351-x}. Epub 2011 Sep 21. PMID: 21935715.}
\item{\strong{Germany}: Greiner W, Claes C, Busschbach JJ, von der Schulenburg JM. Validating the EQ-5D with time trade off for the German population. Eur J Health Econ. 2005 Jun;6(2):124-30. \doi{10.1007/s10198-004-0264-z}. PMID: 19787848.}
\item{\strong{Hungary}: Rencz F, Brodszky V, Gulácsi L, Golicki D, Ruzsa G, Pickard AS, Law EH, Péntek M. Parallel Valuation of the EQ-5D-3L and EQ-5D-5L by Time Trade-Off in Hungary. Value Health. 2020 Sep;23(9):1235-1245. \doi{10.1016/j.jval.2020.03.019}. Epub 2020 Aug 12. PMID: 32940242.}
\item{\strong{Italy}: Scalone L, Cortesi PA, Ciampichini R, Belisari A, D'Angiolella LS, Cesana G, Mantovani LG. Italian population-based values of EQ-5D health states. Value Health. 2013 Jul-Aug;16(5):814-22. \doi{10.1016/j.jval.2013.04.008}. Epub 2013 Jun 19. PMID: 23947975.}
\item{\strong{Japan}: Tsuchiya A, Ikeda S, Ikegami N, Nishimura S, Sakai I, Fukuda T, Hamashima C, Hisashige A, Tamura M. Estimating an EQ-5D population value set: the case of Japan. Health Econ. 2002 Jun;11(4):341-53. \doi{10.1002/hec.673}. PMID: 12007165.}
\item{\strong{Jordan}: Al Rabayah A, Roudijk B, Purba FD, Rencz F, Jaddoua S, Siebert U. Valuation of the EQ-5D-3L in Jordan. Eur J Health Econ. 2024 Sep 3. \doi{10.1007/s10198-024-01712-z}. Epub ahead of print. PMID: 39225720.}
\item{\strong{Netherlands}: Lamers LM, McDonnell J, Stalmeier PF, Krabbe PF, Busschbach JJ. The Dutch tariff: results and arguments for an effective design for national EQ-5D valuation studies. Health Econ. 2006 Oct;15(10):1121-32. \doi{10.1002/hec.1124}. PMID: 16786549.}
\item{\strong{Pakistan}: Malik M, Gu NY, Hussain A, Roudijk B, Purba FD. The EQ-5D-3L Valuation Study in Pakistan. Pharmacoecon Open. 2023 Sep 13. \doi{10.1007/s41669-023-00437-8}. Epub ahead of print. PMID: 37702988.}
\item{\strong{Poland}: Golicki D, Jakubczyk M, Niewada M, Wrona W, Busschbach JJ. Valuation of EQ-5D health states in Poland: first TTO-based social value set in Central and Eastern Europe. Value Health. 2010 Mar-Apr;13(2):289-97. \doi{10.1111/j.1524-4733.2009.00596.x}. Epub 2009 Sep 10. PMID: 19744296.}
\item{\strong{Portugal}: Ferreira LN, Ferreira PL, Pereira LN, Oppe M. The valuation of the EQ-5D in Portugal. Qual Life Res. 2014 Mar;23(2):413-23. \doi{10.1007/s11136-013-0448-z}. Epub 2013 Jun 8. PMID: 23748906.}
\item{\strong{Russia}: Omelyanovskiy V, Musina N, Ratushnyak S, Bezdenezhnykh T, Fediaeva V, Roudijk B, Purba FD. Valuation of the EQ-5D-3L in Russia. Qual Life Res. 2021 Mar 13. \doi{10.1007/s11136-021-02804-6}. Epub ahead of print. PMID: 33713323.}
\item{\strong{Singapore}: Luo N, Wang P, Thumboo J, Lim YW, Vrijhoef HJ. Valuation of EQ-5D-3L health states in Singapore: modeling of time trade-off values for 80 empirically observed health states. Pharmacoeconomics. 2014 May;32(5):495-507. \doi{10.1007/s40273-014-0142-1}. PMID: 24519603.}
\item{\strong{Spain}: Badia X, Roset M, Herdman M, Kind P. A comparison of United Kingdom and Spanish general population time trade-off values for EQ-5D health states. Med Decis Making. 2001 Jan-Feb;21(1):7-16. \doi{10.1177/0272989X0102100102}. PMID: 11206949.}
\item{\strong{South Korea}: Lee YK, Nam HS, Chuang LH, Kim KY, Yang HK, Kwon IS, Kind P, Kweon SS, Kim YT. South Korean time trade-off values for EQ-5D health states: modeling with observed values for 101 health states. Value Health. 2009 Nov-Dec;12(8):1187-93. \doi{10.1111/j.1524-4733.2009.00579.x}. Epub 2009 Jul 29. PMID: 19659703.}
\item{\strong{Sri Lanka}: Kularatna S, Whitty JA, Johnson NW, Jayasinghe R, Scuffham PA. Valuing EQ-5D health states for Sri Lanka. Qual Life Res. 2015 Jul;24(7):1785-93. \doi{10.1007/s11136-014-0906-2}. Epub 2014 Dec 28. PubMed PMID: PMID: 25543271.}
\item{\strong{Sweden}: Burström K, Sun S, Gerdtham UG, Henriksson M, Johannesson M, Levin LÅ, Zethraeus N. Swedish experience-based value sets for EQ-5D health states. Qual Life Res. 2014 Mar;23(2):431-42. \doi{10.1007/s11136-013-0496-4}. PMID: 23975375.}
\item{\strong{Taiwan}: Lee HY, Hung MC, Hu FC, Chang YY, Hsieh CL, Wang JD. Estimating quality weights for EQ-5D (EuroQol-5 dimensions) health states with the time trade-off method in Taiwan. J Formos Med Assoc. 2013 Nov;112(11):699-706. \doi{10.1016/j.jfma.2012.12.015}. Epub 2013 Feb 12. PMID: 24183199.}
\item{\strong{Thailand}: Tongsiri S, Cairns J. Estimating population-based values for EQ-5D health states in Thailand. Value Health. 2011 Dec;14(8):1142-5. \doi{10.1016/j.jval.2011.06.005}. PMID: 22152185.}
\item{\strong{Trinidad and Tobago}: Bailey H, Stolk E, Kind P. Toward Explicit Prioritization for the Caribbean: An EQ-5D Value Set for Trinidad and Tobago. Value Health Reg Issues. 2016 Dec;11:60-67. \doi{10.1016/j.vhri.2016.07.010}. PMID: 27986200.}
\item{\strong{Tunisia}: Chemli J, Drira C, Felfel H, Roudijk B, Al Sayah F, Kouki M, Kooli A, Razgallah Khrouf M. Valuing health-related quality of life using a hybrid approach: Tunisian value set for the EQ-5D-3L. Qual Life Res. 2021 Jan 14. \doi{10.1007/s11136-020-02730-z}. Epub ahead of print. PMID: 33447958.}
\item{\strong{UK}: Dolan P. Modeling valuations for EuroQol health states. Med Care. 1997 Nov;35(11):1095-108. \doi{10.1097/00005650-199711000-00002}. PMID: 9366889.}
\item{\strong{USA}: Shaw JW, Johnson JA, Coons SJ. US valuation of the EQ-5D health states: development and testing of the D1 valuation model. Med Care. 2005 Mar;43(3):203-20. \doi{10.1097/00005650-200503000-00003}. PMID: 15725977.}
\item{\strong{Zimbabwe}: Jelsma J, Hansen K, De Weerdt W, De Cock P, Kind P. How do Zimbabweans value health states? Popul Health Metr. 2003 Dec 16;1(1):11. \doi{10.1186/1478-7954-1-11}. PMID: 14678566.}
}
}
\usage{
TTO
}
\description{
Coefficients for the estimation of the EQ-5D-3L index values based on TTO valuation studies
  for Argentina, Australia, Brazil, Canada, Chile, China, Denmark, Ecuador, 
  France, Germany, Hungary, Italy, Japan, Netherlands, Pakistan, Poland, Portugal, 
  Russia, Singapore, SouthKorea, Spain, SriLanka, Sweden, Taiwan, Thailand, 
  Trinidad and Tobago, Tunisia, UK, USA and Zimbabwe.
}
\keyword{datasets}
