% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VAS}
\alias{VAS}
\title{EQ-5D-3L VAS value set data}
\format{An object of class \code{data.frame} with 13 rows and 9 columns.}
\source{
Oppe, M., Szende, A., & de Charro, F. (2007), Comparative review of Visual Analogue Scale value sets.
  In Szende, A., Oppe, M., & Devlin, N. (Ed.), EQ-5D Value Sets: Inventory, Comparative Review
  and User Guide (pp. 37-38). Dordrecht, The Netherlands: Springer.
}
\usage{
VAS
}
\description{
Coefficients for the estimation of the EQ-5D-3L index values based on VAS valuation studies
 for Belgium, Denmark, Europe, Finland, Germany, New Zealand, Slovenia, Spain and UK.
}
\keyword{datasets}
